Move SWE backends into package folders
