+++ verifiers/envs/experimental/composable/tasksets/swe/scale_swe/__init__.py
@@ -1,0 +1,3 @@
+from .taskset import ScaleSWERubric, ScaleSWETaskSet
+
+__all__ = ["ScaleSWERubric", "ScaleSWETaskSet"]
