+++ verifiers/envs/experimental/composable/tasksets/swe/shared/__init__.py
@@ -1,0 +1,3 @@
+from .test_patch import revert_and_reapply_test_patch
+
+__all__ = ["revert_and_reapply_test_patch"]
