+++ verifiers/envs/experimental/composable/tasksets/swe/swe_bench/__init__.py
@@ -1,0 +1,3 @@
+from .taskset import SWEBenchRubric, SWEBenchTaskSet
+
+__all__ = ["SWEBenchRubric", "SWEBenchTaskSet"]
