+++ verifiers/envs/experimental/composable/tasksets/swe/swe_lego/__init__.py
@@ -1,0 +1,3 @@
+from .taskset import SWELegoRubric, SWELegoTaskSet
+
+__all__ = ["SWELegoRubric", "SWELegoTaskSet"]
