+++ verifiers/envs/experimental/composable/tasksets/swe/swe_rebench_v2/taskset.py
@@ -35,7 +35,7 @@ import verifiers as vf
 from datasets import load_dataset
 from verifiers.envs.experimental.composable import SandboxSpec, SandboxTaskSet
 
-from verifiers.envs.experimental.composable.tasksets.swe._test_patch import (
+from verifiers.envs.experimental.composable.tasksets.swe.shared.test_patch import (
     revert_and_reapply_test_patch,
 )
