+++ verifiers/envs/experimental/composable/tasksets/swe/swe_smith/__init__.py
@@ -1,0 +1,3 @@
+from .taskset import SWESmithRubric, SWESmithTaskSet
+
+__all__ = ["SWESmithRubric", "SWESmithTaskSet"]
